+++ CHANGELOG.rst
@@ -18,7 +18,7 @@ Changelog
   :class:`~cryptography.hazmat.primitives.asymmetric.dsa.DSAPrivateKeyWithSerialization`,
   and
   :class:`~cryptography.hazmat.primitives.asymmetric.ec.EllipticCurvePrivateKeyWithSerialization`.
-* Add support for parsing X.509 CSRs (certificate signing requests) with
+* Add support for parsing X.509 certificate signing requests (CSRs) with
   :func:`~cryptography.x509.load_pem_x509_csr`.
 
 0.8.1 - 2015-03-20